Fixed bug in setup_arduino_libraries
If the first source file doesn't contain includes to all required arduino
libraries, the original search code failed. " around the SRCS variable were
missing and cmake interpreted the sequence of source files as a sequence of
parameters. That means that the actual library search function only got one
source file, the first one and was performing the search only there.