Added support for /wall_clock topic
Since it is really hard to get the clock set correctly on an android device, we
need a different mechanism for getting time-stamps synchronized with a PC.
Added setting of time through the topic /wall_clock. Whenever a time is
published on this topic, it is used as the device's current time and the
arduino is notified to update its time, too.