Fixed sending of time in rosserial_java
We must not send the new time to the arduino as soon as we received it. The
problem is that rosserial on the micro-controller tries to calibrate
communication latency. If we send a time message outside of a time request, the
micro-controller's clock will be wrong. Since the micro-controller requests a
new time each second anyway, we don't need to send it.